% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBarplot.R
\name{CreateBarplot}
\alias{CreateBarplot}
\title{Create a barplot of gene expression rates}
\usage{
CreateBarplot(disease, output_directory, file_prefix, simulated_matrix,
  tcga_matrix_tumor, top_DEG, sample_number)
}
\arguments{
\item{disease}{String; cancer name}

\item{output_directory}{String; name of the output directory}

\item{file_prefix}{String; name of the results output file}

\item{sample_number}{Integer; the number of samples which are simulated}
}
\description{
Create a barplot of gene expression rates
}
