% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMAF.R
\name{GenerateMAF}
\alias{GenerateMAF}
\title{Generate MAF file out of QTL data (from csv)}
\usage{
GenerateMAF(threshold_eQTls, tumor_sample_barcodes, output_directory,
  file_prefix, disease, tcga_matrix_normal, tcga_matrix_tumor)
}
\arguments{
\item{threshold_eQTls}{Int; Threshold of the number of eQTLs used}

\item{tumor_sample_barcodes}{}
}
\value{
Matrix; Simulated Matrix
}
\description{
Generate MAF file out of QTL data (from csv)
}
