% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesFromPathway.R
\name{GenesFromPathway}
\alias{GenesFromPathway}
\title{Get gene expression relations between genes from a specific pathway}
\usage{
GenesFromPathway(species, pID)
}
\arguments{
\item{species}{}

\item{pID}{String; id of the KEGG pathway}
}
\description{
Get gene expression relations between genes from a KEGG pathway and write them in input file, using pathway specific kgml file of KEGGREST database
}
