% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCancerProject.R
\name{GetCancerProject}
\alias{GetCancerProject}
\title{Get Cancer project from TCGA}
\usage{
GetCancerProject(disease)
}
\arguments{
\item{disease}{}
}
\value{
String; code for the TCGA Cancer Project
}
\description{
Get Cancer project from TCGA
}
