% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInputData.R
\name{GetInputData}
\alias{GetInputData}
\title{Read given gene symbols out of txt file}
\usage{
GetInputData()

GetInputData()
}
\value{
List of all gene symbols with variations and dictionary of gene symbols and the specific influence on their expression
}
\description{
Reads data out of txt file and puts it into a list, containing gene symbols of genes with variation, and a dictionary,
containing gene symbols of genes mapped to their influence on the expression value.
Every row of the txt file has to be formatted this way:
GeneSymbolOfGeneWithVariation[String] GeneSymbolOfGeneWithChangedExpressionValue[String] Influence[Characters --,-,++,+]
}
\examples{
GetInputData()
}
