% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPathwayID.R
\name{GetPathwayID}
\alias{GetPathwayID}
\title{Mapping of disease name to pathway name}
\usage{
GetPathwayID(disease)
}
\arguments{
\item{disease}{String; name of a disease to get a KEGG pathway for}
}
\value{
String; id of the KEGG pathway
}
\description{
Mapping of disease name to pathway name
}
