% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopGenesFromOpentargets.R
\name{GetTopGenesFromOpentargets}
\alias{GetTopGenesFromOpentargets}
\title{GetTopGenesFromOpentargets}
\usage{
GetTopGenesFromOpentargets(efo_code, top_DEG_number)
}
\value{
top genes
}
\description{
Access opentargets.org REST style API and download data
}
