% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTCGAMatrix.R
\name{LoadTCGAMatrix}
\alias{LoadTCGAMatrix}
\title{Load the TCGA Matrix of specified disease}
\usage{
LoadTCGAMatrix(disease, sample_type, sample_number)
}
\arguments{
\item{sample_type}{}
}
\value{
tcga_matrix
}
\description{
Load the TCGA Matrix of specified disease
}
