% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDEG.R
\name{TopDEG}
\alias{TopDEG}
\title{Return top differentially expressed genes}
\usage{
TopDEG(results_DEG, top_DEG_number)
}
\arguments{
\item{results_DEG}{Matrix; differentaial analysis results}

\item{top_DEG_number}{Integer; number of top differential expressed genes}
}
\description{
Return top differentially expressed genes
}
