% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA}
\alias{DEA}
\title{DEA script}
\usage{
DEA(disease, sample_number, output_directory, count_matrix_control,
  count_matrix_test, description)
}
\arguments{
\item{description}{}
}
\description{
DEA script
}
