% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omicsimulator.R
\name{Omicsimulator}
\alias{Omicsimulator}
\title{Main function which calls all relevant to simulate the influence of variations on gene expression}
\usage{
Omicsimulator(disease, sample_number, top_DEG_number, output_directory,
  file_prefix)

Omicsimulator(disease, sample_number, top_DEG_number, output_directory,
  file_prefix)
}
\arguments{
\item{disease}{String; name of a disease to get a KEGG pathway for}

\item{sample_number}{Integer; the number of samples which are simulated}

\item{top_DEG_number}{Integer; number of top differential expressed genes}

\item{output_directory}{String; directory of the output files}

\item{file_prefix}{String; name of the results output file}
}
\value{
List; List containing the DEA results of both normal-tumor and normal-simulated
}
\description{
Main function which calls all relevant to simulate the influence of variations on gene expression
}
\examples{
Omicsimulator(disease = "Breast cancer", sample_number = 10, top_DEG_number = 100)
}
